move formating to separate file, setup __init__ for imports
